Renames preadj package in bfd-server-war.
