+++ apps/bfd-server/bfd-server-war/src/main/java/gov/cms/bfd/server/war/SpringConfiguration.java
@@ -17,8 +17,8 @@ import gov.cms.bfd.model.rif.schema.DatabaseTestUtils.DataSourceComponents;
 import gov.cms.bfd.server.war.r4.providers.R4CoverageResourceProvider;
 import gov.cms.bfd.server.war.r4.providers.R4ExplanationOfBenefitResourceProvider;
 import gov.cms.bfd.server.war.r4.providers.R4PatientResourceProvider;
-import gov.cms.bfd.server.war.r4.providers.preadj.R4ClaimResourceProvider;
-import gov.cms.bfd.server.war.r4.providers.preadj.R4ClaimResponseResourceProvider;
+import gov.cms.bfd.server.war.r4.providers.partadj.R4ClaimResourceProvider;
+import gov.cms.bfd.server.war.r4.providers.partadj.R4ClaimResponseResourceProvider;
 import gov.cms.bfd.server.war.stu3.providers.CoverageResourceProvider;
 import gov.cms.bfd.server.war.stu3.providers.ExplanationOfBenefitResourceProvider;
 import gov.cms.bfd.server.war.stu3.providers.PatientResourceProvider;
