+++ apps/bfd-server/bfd-server-war/src/main/java/gov/cms/bfd/server/war/r4/providers/partadj/FissClaimResponseTransformerV2.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.server.war.r4.providers.preadj;
+package gov.cms.bfd.server.war.r4.providers.partadj;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.parser.IParser;
@@ -9,7 +9,7 @@ import gov.cms.bfd.model.rda.PartAdjFissClaim;
 import gov.cms.bfd.model.rda.PartAdjFissPayer;
 import gov.cms.bfd.server.war.commons.BBCodingSystems;
 import gov.cms.bfd.server.war.commons.carin.C4BBIdentifierType;
-import gov.cms.bfd.server.war.r4.providers.preadj.common.AbstractTransformerV2;
+import gov.cms.bfd.server.war.r4.providers.partadj.common.AbstractTransformerV2;
 import gov.cms.bfd.sharedutils.exceptions.BadCodeMonkeyException;
 import java.util.ArrayList;
 import java.util.Date;
